build: fix optimize images. Fixes #21
Image optimization process was failing in CI because 'public/optimized' folder didn't exist

+++ optimize-images.js
@@ -6,11 +6,15 @@ const fs = require('fs');
 const path = require('path');
 
 (async () => {
-  const currentFilenames = fs.readdirSync('public/optimized');
-  currentFilenames.forEach(filename => {
-    fs.unlinkSync(path.join('public/optimized', filename));
-  });
+  // Clear previous optimized resources if necessary
+  if (fs.existsSync('public/optimized')) {
+    const currentFilenames = fs.readdirSync('public/optimized');
+    currentFilenames.forEach(filename => {
+      fs.unlinkSync(path.join('public/optimized', filename));
+    });
+  }
 
+  // Optimize images
   const files = await imagemin(['src/public/*.png'], {
     destination: 'public/optimized',
     plugins: [
@@ -21,6 +25,7 @@ const path = require('path');
     ],
   });
 
+  // Hash images
   const filenameToHash = {};
   files.forEach(f => {
     const hash = crypto.createHash('sha256');
@@ -34,11 +39,13 @@ const path = require('path');
       .substring(0, 8)}${pathParts.ext}`;
   });
 
+  // Dump hashes to `img-hash.json` for app consumption
   fs.writeFileSync(
     'public/optimized/img-hash.json',
     JSON.stringify(filenameToHash),
   );
 
+  // Include hash in all file names to boost caching
   Object.entries(filenameToHash).forEach(([filename, hashedFilename]) => {
     const currentFilename = path.join('public/optimized', filename);
     const nextFilename = path.join('public/optimized', hashedFilename);
